+++ browser-test/src/admin/admin_application_download.test.ts
@@ -381,9 +381,9 @@ test.describe('csv json pdf download test- two applications', () => {
       await loginAsProgramAdmin(page)
       await adminPrograms.viewApplications(programName)
       await adminPrograms.filterProgramApplications({searchFragment: 'SARA'})
-      await adminPrograms.viewApplicationForApplicant('smith, sarah')
+      await adminPrograms.viewApplicationForApplicantForBulkStatus('smith, sarah')
 
-      const pdfFile = await adminPrograms.getApplicationPdf()
+      const pdfFile = await adminPrograms.getApplicationPdfForBulkStatus()
       expect(pdfFile.length).toBeGreaterThan(1)
       await logout(page)
     })
