+++ browser-test/src/applicant/application_previous_version.test.ts
@@ -6,6 +6,7 @@ import {
   loginAsTestUser,
   logout,
   testUserDisplayName,
+  enableFeatureFlag,
 } from '../support'
 
 test.describe('view an application in an older version', () => {
@@ -14,6 +15,7 @@ test.describe('view an application in an older version', () => {
     adminQuestions,
     adminPrograms,
   }) => {
+    await enableFeatureFlag(page, 'bulk_status_update_enabled')
     await loginAsAdmin(page)
 
     // Create a program with one question
@@ -41,13 +43,15 @@ test.describe('view an application in an older version', () => {
 
     // See the application in admin page
     await adminPrograms.viewApplications(programName)
-    await adminPrograms.viewApplicationForApplicant(testUserDisplayName())
-    await adminPrograms.expectApplicationAnswers(
+    await adminPrograms.viewApplicationForApplicantForBulkStatus(
+      testUserDisplayName(),
+    )
+    await adminPrograms.expectApplicationAnswersForBulkStatus(
       'Screen 1',
       questionName,
       'some text',
     )
-
+    await page.getByRole('link', {name: 'Back'}).click()
     await logout(page)
     await loginAsAdmin(page)
 
@@ -60,8 +64,10 @@ test.describe('view an application in an older version', () => {
 
     // See the application in admin page in the old version
     await adminPrograms.viewApplications(programName)
-    await adminPrograms.viewApplicationForApplicant(testUserDisplayName())
-    await adminPrograms.expectApplicationAnswers(
+    await adminPrograms.viewApplicationForApplicantForBulkStatus(
+      testUserDisplayName(),
+    )
+    await adminPrograms.expectApplicationAnswersForBulkStatus(
       'Screen 1',
       questionName,
       'some text',
